Fix logging of requireSameDimensions setting
Ensure that the value of requireSameDimensions reported to the user is
the same value used in the test of requireSameDimensions in
`compare-resemble.js`.

In cases where the user has not configured requireSameDimensions, its
value will be `undefined`. This used to cause BackstopJS to log that
requireSameDimensions was false; however the BackstopJS documentation
says that the default value of requireSameDimensions is `true`, and the
test of requireSameDimensions in `compare-resemble.js` asserts that it
must be `false` to be considered falsey.

This commit updates the logging of requireSameDimensions when reporting
a mismatch, so that it is clear to the user that requireSameDimensions
is being treated as `true`.